+++ test/quack_db/protocol/scalar_decoding_test.exs
@@ -12,7 +12,12 @@ defmodule QuackDB.Protocol.ScalarDecodingTest do
         {:varchar, :varchar, ["duck", nil, "quack"]},
         {:blob, :varchar, [<<1, 2>>, <<>>, nil]},
         {:date, :int32, [0, 1, -1]},
+        {:time_ns, :int64, [1_234_567_890, 0, 42]},
+        {:time_tz, :int64, [16_777_216_057_599, 16_777_216_053_999, 16_777_216_061_199]},
         {:timestamp, :int64, [0, 1_000_000, -1_000_000]},
+        {:timestamp_ns, :int64, [1_234_567_890, 0, -1_234_567_890]},
+        {:interval, :interval, [{1, 2, 3}, {0, 1, 1000}, nil]},
+        {:bignum, :varchar, [<<128, 0, 1, 1>>, <<127, 255, 254, 254>>, <<128, 0, 2, 1, 0>>]},
         {:decimal, :int64, [12345, -500, 0]}
       ])
       |> IO.iodata_to_binary()
@@ -27,7 +32,12 @@ defmodule QuackDB.Protocol.ScalarDecodingTest do
                "duck",
                <<1, 2>>,
                ~D[1970-01-01],
+               QuackDB.NanosecondTime.new(1_234_567_890),
+               QuackDB.TimeWithTimeZone.new(~T[00:00:01.000000], 0),
                ~N[1970-01-01 00:00:00.000000],
+               QuackDB.NanosecondTimestamp.new(1_234_567_890),
+               QuackDB.Interval.new(1, 2, 3),
+               1,
                Decimal.new(1, 12345, -2)
              ],
              [
@@ -37,7 +47,12 @@ defmodule QuackDB.Protocol.ScalarDecodingTest do
                nil,
                <<>>,
                ~D[1970-01-02],
+               QuackDB.NanosecondTime.new(0),
+               QuackDB.TimeWithTimeZone.new(~T[00:00:01.000000], 3600),
                ~N[1970-01-01 00:00:01.000000],
+               QuackDB.NanosecondTimestamp.new(0),
+               QuackDB.Interval.new(0, 1, 1000),
+               -1,
                Decimal.new(-1, 500, -2)
              ],
              [
@@ -47,7 +62,12 @@ defmodule QuackDB.Protocol.ScalarDecodingTest do
                "quack",
                nil,
                ~D[1969-12-31],
+               QuackDB.NanosecondTime.new(42),
+               QuackDB.TimeWithTimeZone.new(~T[00:00:01.000000], -3600),
                ~N[1969-12-31 23:59:59.000000],
+               QuackDB.NanosecondTimestamp.new(-1_234_567_890),
+               nil,
+               256,
                Decimal.new(1, 0, -2)
              ]
            ]